+++ .gitignore
@@ -1,42 +1,91 @@
-﻿/.idea/.idea.TelegramBotCore/
-/src/TelegramBotCore.API/bin/Debug/netcoreapp3.1/
-/src/TelegramBotCore.API/obj/
-/src/TelegramBotCore.API/appsettings.Development.json
-/src/TelegramBotCore.Core/obj/
-/src/BotCore.API/bin/Debug/netcoreapp3.1/Properties/launchSettings.json
-/src/BotCore.API/bin/Debug/netcoreapp3.1/appsettings.Development.json
-/src/BotCore.API/bin/Debug/netcoreapp3.1/appsettings.json
-/src/BotCore.API/bin/Debug/netcoreapp3.1/TelegramBotCore.API.deps.json
-/src/BotCore.API/bin/Debug/netcoreapp3.1/TelegramBotCore.API.runtimeconfig.dev.json
-/src/BotCore.API/bin/Debug/netcoreapp3.1/TelegramBotCore.API.runtimeconfig.json
-/src/BotCore.API/obj/Debug/netcoreapp3.1/staticwebassets/TelegramBotCore.API.StaticWebAssets.xml
-/src/BotCore.API/obj/Debug/netcoreapp3.1/.NETCoreApp,Version=v3.1.AssemblyAttributes.cs
-/src/BotCore.API/obj/Debug/netcoreapp3.1/BotCore.API.AssemblyInfo.cs
-/src/BotCore.API/obj/Debug/netcoreapp3.1/TelegramBotCore.API.AssemblyInfo.cs
-/src/BotCore.API/obj/Debug/netcoreapp3.1/TelegramBotCore.API.csproj.FileListAbsolute.txt
-/src/BotCore.API/obj/BotCore.API.csproj.nuget.dgspec.json
-/src/BotCore.API/obj/BotCore.API.csproj.nuget.g.props
-/src/BotCore.API/obj/BotCore.API.csproj.nuget.g.targets
-/src/BotCore.API/obj/project.assets.json
-/src/BotCore.API/obj/project.packagespec.json
-/src/BotCore.API/obj/rider.project.restore.info
-/src/BotCore.API/obj/TelegramBotCore.API.csproj.nuget.dgspec.json
-/src/BotCore.API/obj/TelegramBotCore.API.csproj.nuget.g.props
-/src/BotCore.API/obj/TelegramBotCore.API.csproj.nuget.g.targets
-/src/BotCore.API/appsettings.Development.json
-/src/BotCore.Core/obj/Debug/netcoreapp3.0/.NETCoreApp,Version=v3.0.AssemblyAttributes.cs
-/src/BotCore.Core/obj/Debug/netcoreapp3.0/BotCore.Core.AssemblyInfo.cs
-/src/BotCore.Core/obj/Debug/netcoreapp3.0/TelegramBotCore.Core.AssemblyInfo.cs
-/src/BotCore.Core/obj/BotCore.Core.csproj.nuget.dgspec.json
-/src/BotCore.Core/obj/BotCore.Core.csproj.nuget.g.props
-/src/BotCore.Core/obj/BotCore.Core.csproj.nuget.g.targets
-/src/BotCore.Core/obj/project.assets.json
-/src/BotCore.Core/obj/project.packagespec.json
-/src/BotCore.Core/obj/rider.project.restore.info
-/src/BotCore.Core/obj/TelegramBotCore.Core.csproj.nuget.dgspec.json
-/src/BotCore.Core/obj/TelegramBotCore.Core.csproj.nuget.g.props
-/src/BotCore.Core/obj/TelegramBotCore.Core.csproj.nuget.g.targets
-/.idea/.idea.BotCore/.idea/encodings.xml
-/.idea/.idea.BotCore/.idea/workspace.xml
+﻿# Created by .ignore support plugin (hsz.mobi)
+### JetBrains template
+# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
+# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
+
+# User-specific stuff
+.idea/**/workspace.xml
+.idea/**/tasks.xml
+.idea/**/usage.statistics.xml
+.idea/**/dictionaries
+.idea/**/shelf
+
+# Generated files
+.idea/**/contentModel.xml
+
+# Sensitive or high-churn files
+.idea/**/dataSources/
+.idea/**/dataSources.ids
+.idea/**/dataSources.local.xml
+.idea/**/sqlDataSources.xml
+.idea/**/dynamic.xml
+.idea/**/uiDesigner.xml
+.idea/**/dbnavigator.xml
+
+# Gradle
+.idea/**/gradle.xml
+.idea/**/libraries
+
+# Gradle and Maven with auto-import
+# When using Gradle or Maven with auto-import, you should exclude module files,
+# since they will be recreated, and may cause churn.  Uncomment if using
+# auto-import.
+# .idea/artifacts
+# .idea/compiler.xml
+# .idea/jarRepositories.xml
+# .idea/modules.xml
+# .idea/*.iml
+# .idea/modules
+# *.iml
+# *.ipr
+
+# CMake
+cmake-build-*/
+
+# Mongo Explorer plugin
+.idea/**/mongoSettings.xml
+
+# File-based project format
+*.iws
+
+# IntelliJ
+out/
+
+# mpeltonen/sbt-idea plugin
+.idea_modules/
+
+# JIRA plugin
+atlassian-ide-plugin.xml
+
+# Cursive Clojure plugin
+.idea/replstate.xml
+
+# Crashlytics plugin (for Android Studio and IntelliJ)
+com_crashlytics_export_strings.xml
+crashlytics.properties
+crashlytics-build.properties
+fabric.properties
+
+# Editor-based Rest Client
+.idea/httpRequests
+
+# Android studio 3.1+ serialized cache file
+.idea/caches/build_file_checksums.ser
+
+### VisualStudioCode template
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+*.code-workspace
+
+# Local History for Visual Studio Code
+.history/
+
+/src/BotCore.API/
+/src/BotCore.Core/
+/src/BotCore.Telegram/bin/Debug/netcoreapp3.0/
+/src/BotCore.Telegram/obj/
 /BotCore.Telegram/obj/
-/.idea/.idea.BotCore/
+/.idea/
