+++ BotCore.sln
@@ -4,7 +4,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BotCore.API", "src\BotCore.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BotCore.Core", "src\BotCore.Core\BotCore.Core.csproj", "{B2611AD5-F0A3-43D2-B853-BAED1EFABC40}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BotCore.Telegram", "BotCore.Telegram\BotCore.Telegram.csproj", "{154B99EC-9358-4774-AF90-2517B0DF256D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BotCore.Telegram", "src\BotCore.Telegram\BotCore.Telegram.csproj", "{5626168A-DEF0-4AD8-A8EE-1C74C9285780}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,9 +20,9 @@ Global
 		{B2611AD5-F0A3-43D2-B853-BAED1EFABC40}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2611AD5-F0A3-43D2-B853-BAED1EFABC40}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2611AD5-F0A3-43D2-B853-BAED1EFABC40}.Release|Any CPU.Build.0 = Release|Any CPU
-		{154B99EC-9358-4774-AF90-2517B0DF256D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{154B99EC-9358-4774-AF90-2517B0DF256D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{154B99EC-9358-4774-AF90-2517B0DF256D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{154B99EC-9358-4774-AF90-2517B0DF256D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5626168A-DEF0-4AD8-A8EE-1C74C9285780}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5626168A-DEF0-4AD8-A8EE-1C74C9285780}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5626168A-DEF0-4AD8-A8EE-1C74C9285780}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5626168A-DEF0-4AD8-A8EE-1C74C9285780}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
