+++ src/BotCore.API/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace TelegramBotCore.API
+namespace BotCore.API
 {
     public class Program
     {
