+++ src/BotCore.Core/Interfaces/IAction.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace TelegramBotCore.Core.Interfaces
+namespace BotCore.Core.Interfaces
 {
     public interface IAction
     {
